refactor: hono api
