+++ src/app/api/[...route]/route.ts
@@ -1,8 +1,8 @@
 import { handle } from "hono/vercel";
-import oauthRoutes from "@/routes/oauth";
-import authRoutes from "@/routes/auth";
-import slackRoutes from "@/routes/slack";
-import healthRoutes from "@/routes/health";
+import oauthRoutes from "@/handlers/api/oauth";
+import authRoutes from "@/handlers/api/auth";
+import slackRoutes from "@/handlers/api/slack";
+import healthRoutes from "@/handlers/api/health";
 import { createHonoApp } from "../../create-app";
 
 const app = createHonoApp().basePath("/api");
